renamed `spec_binder.rs` to `partition_binder.rs`
