+++ crates/iceberg-ext/src/spec/mod.rs
@@ -6,6 +6,6 @@ pub use iceberg::spec::{
     UnboundPartitionField, UnboundPartitionSpec, ViewMetadata, ViewVersion,
 };
 
-mod spec_binder;
+mod partition_binder;
 mod table_metadata;
 pub use table_metadata::TableMetadataBuilder;
